+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # dionysos ChangeLog
 
+# Version 0.2.0 (IN PROGRESS)
+
+- Manage volume
+- [GH-2][]: MPlayer backend
+
+
 # Version 0.1.0 (15/01/2015)
 
 - Init files mode (directory listing)
@@ -8,9 +14,9 @@
 - Init project
 
 
-[TravisCI]: https://travis-ci.org/nlamirault/marcopolo
-[Drone.io]: https://drone.io/github.com/nlamirault/marcopolo
-[overseer]: https://github.com/tonini/overseer.el
+[TravisCI]: https://travis-ci.org/nlamirault/dionysos
+[Drone.io]: https://drone.io/github.com/nlamirault/dionysos
 
 
 [GH-1]: https://github.com/nlamirault/dionysos/issues/1
+[GH-2]: https://github.com/nlamirault/dionysos/issues/2